Refactoring account api
