+++ internal/controller/controller.go
@@ -1,9 +1,6 @@
 package controller
 
 import (
-	"net/http"
-	"time"
-
 	"github.com/gin-gonic/gin"
 
 	ginSwagger "github.com/swaggo/gin-swagger"
@@ -33,32 +30,11 @@ func (h *Controller) InitRoutes() *gin.Engine {
 
 	api := router.Group("/api")
 	{
-		api.POST("/auth/login", h.getToken)
-		api.POST("/auth/registration", h.registerUser)
-		api.POST("/auth/activate", h.activateAccount)
 		api.GET("/user", h.getUser)
+		h.initAccountRoutes(api)
 		h.initChallengesRoutes(api)
 		h.initTimelineRoutes(api)
 	}
 
 	return router
-}
-
-// TODO: перенести в другой файл
-func (h *Controller) getToken(c *gin.Context) {
-	var input map[string]interface{}
-	if err := c.BindJSON(&input); err != nil {
-		c.AbortWithStatusJSON(http.StatusBadRequest, err)
-		return
-	}
-
-	token, err := h.tokenManager.NewJWT(input["id"].(string), time.Hour)
-	if err != nil {
-		c.AbortWithStatusJSON(http.StatusInternalServerError, createMessage(err.Error()))
-		return
-	}
-
-	c.JSON(http.StatusOK, map[string]interface{}{
-		"token": token,
-	})
 }
